figurabackend/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
defusedxml==0.5.0
Django==2.0.8
drf-haystack==1.8.4
django-allauth==0.37.1
django-cors-headers==2.4.0
django-hashid-field==2.1.1
django-ordered-model==2.1.0
django-resized==0.3.9
django-rest-auth==0.9.3
django-simple-history==2.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-serializer-extensions==0.6.0
djangorestframework-simplejwt==3.2.3
docopt==0.6.2
dry-rest-permissions==0.1.10
hashids==1.2.0
idna==2.7
lxml==4.2.4
oauthlib==2.1.0
Pillow==5.2.0
pipreqs==0.4.9
pkg-resources==0.0.0
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.6.4
pyreqs==0.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.1
requests-oauthlib==1.0.0
sh==1.12.14
six==1.11.0
urllib3==1.23
xmljson==0.1.9
yarg==0.1.9
django-extra-fields==1.0.0
uvicorn==0.3.6
gunicorn==19.9.0
channels==2.1.3
channels-redis==2.3.0
pysolr==3.8.1
django-cleanup==2.1.0
django-guardian==1.4.9
